+++ CHANGES.md
@@ -4,6 +4,8 @@ CytoMeth - Changes
 * reshape2 fix
 * updated Dockerfile and install.packages.R files
 * autodetect total memory size and cpu cores and fix the config if needed
+* reference data directory change
+* manual updated 
 
 ## Version 0.9.14
 * new plot functions plotSitesCpG and plotSitesNonCpG allow for any min_coverage